% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getEBEstimates}
\alias{getEBEstimates}
\title{Extract Empirical Bayes Estimates from RebelFitObj}
\usage{
getEBEstimates(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelFit}
}
\description{
Extract Empirical Bayes Estimates from RebelFitObj
}
