% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getEBParameters}
\alias{getEBParameters}
\title{Extract Empirical Bayes Parameters from RebelFitObj}
\usage{
getEBParameters(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelFit}
}
\description{
Extract Empirical Bayes Parameters from RebelFitObj
}
