% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getGeneNames}
\alias{getGeneNames}
\title{Extract Gene Names from RebelFitObj}
\usage{
getGeneNames(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelFit}
}
\description{
Extract Gene Names from RebelFitObj
}
