% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getModelMatrix}
\alias{getModelMatrix}
\title{Extract Model Matrix from RebelFitObj}
\usage{
getModelMatrix(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelFit}
}
\description{
Extract Model Matrix from RebelFitObj
}
