% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{getOriginalFitVar}
\alias{getOriginalFitVar}
\title{Extract Original Variance Estimates from RebelFitObj}
\usage{
getOriginalFitVar(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelFit}
}
\description{
Extract Original Variance Estimates from RebelFitObj
}
