% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBEL-package.R
\docType{package}
\name{REBEL-package}
\alias{REBEL}
\alias{REBEL-package}
\title{REBEL: Repeated Measures Empirical Bayes Differential Expression Using LMM}
\description{
Simulate data from repeated measures scRNA-seq data and analyze differential expression between groups/timepoints in individual cell types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewynn610/REBEL}
  \item \url{https://ewynn610.github.io/REBEL/}
  \item Report bugs at \url{https://github.com/ewynn610/REBEL/issues}
}

}
\keyword{internal}
