% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebelAggregate.R
\name{rebelAggregate}
\alias{rebelAggregate}
\title{Aggregate data for pseudo-bulk analysis}
\usage{
rebelAggregate(
  object,
  countsAssay = "counts",
  counts = NULL,
  colData = NULL,
  sampleVariable
)
}
\arguments{
\item{object}{\code{\link[SingleCellExperiment]{SingleCellExperiment}} object
containing count values in one of the object assays and cell level meta data
in the colData slot.}

\item{countsAssay}{String specifying assay in object which holds count values.}

\item{counts}{Optional matrix-like object containing counts. Used in
conjunction with \code{colData} argument in place of object. Matrix should
contain genes in the rows and cells in the columns.}

\item{colData}{Optional dataframe object containing cell level meta data
which is used with counts argument in place of object. Row names of dataframe
 should match column names of \code{normalizedCounts}.}

\item{sampleVariable}{String denoting name of sample identifier variable in
the \code{colData} which counts should be aggregated over.}
}
\value{
\code{\link[SummarizedExperiment]{SummarizedExperiment}} and includes
pseudo-bulk counts in the \code{counts} assay. All colData variables that do not
vary within the \code{sampleVariable} are also included in the \code{colData}
slot.
}
\description{
Aggregate data for pseudo-bulk analysis
}
\examples{
## Read in single cell data
data("RecAM_sim_sce")

## Aggregate into pseudo-bulk data
pb_dat <- rebelAggregate(object=RecAM_sim_sce, countsAssay="counts", sampleVariable="sampleID")
}
\author{
Elizabeth Wynn
}
