% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebelEmpBayes.R
\name{rebelEmpBayes}
\alias{rebelEmpBayes}
\title{Run Empirical Bayes procedure on linear mixed models}
\usage{
rebelEmpBayes(RebelFitObj)
}
\arguments{
\item{RebelFitObj}{Object created by running rebelLMM}
}
\value{
A \code{\link[REBEL]{RebelFit-class} object}.
}
\description{
Run Empirical Bayes procedure on linear mixed models
}
