% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RebelFit-class}
\alias{RebelFit-class}
\title{The RebelFit class}
\description{
The RebelFit class
}
\section{Slots}{

\describe{
\item{\code{geneNames}}{Vector containing gene names taken from the row names of the
\code{object} or \code{normalizedCounts} provided in the \code{rebelFit}
function call}

\item{\code{fits}}{List of \code{\link[lmerTest]{lmerModLmerTest}} fit objects with one object for each gene.
Not included if \code{outputFits=FALSE}.}

\item{\code{modelMatrix}}{Design matrix for the model fixed effects.}

\item{\code{coefficients}}{Dataframe of model coefficients with columns representing
the model fixed effects and each row representing an individual gene.}

\item{\code{originalFitVar}}{Dataframe of residual and random effect variance values
from original model fits before empirical Bayes estimation. A logical value
indicating whether the original fit was singular or not is also included.}

\item{\code{miscFitInfo}}{Miscellaneous information from model fits that is needed
for hypothesis testing.}

\item{\code{sampleVariable}}{String denoting name of sample identifier variable in
the \code{colData} which was used as the sample-level random effect in the LMM
model. For pseudo-bulk data where no sample-level random effect is used,
this value is \code{NULL}.}

\item{\code{subjectVariable}}{String denoting name of subject identifier variable in
the \code{colData} which was used as the subject-level random effect in the LMM
model.}

\item{\code{pseudoBulk}}{Logical value indicating whether a pseudo-bulk or cell-level
analysis is being performed.}

\item{\code{EBEstimates}}{Dataframe with the empirical Bayes residual variance and
random intercept variance for each gene.}

\item{\code{EBParameters}}{Dataframe with the parameters used to get empirical Bayes
estimates.}

\item{\code{vcovBetaList}}{List of variance/covariance matrices with one matrix per
gene.}

\item{\code{vcovBetaAdjList}}{List of adjusted variance/covariance matrices for each
gene calculated using the Kenward-Roger's method. Additional parameters used
for the Kenward-Roger's degrees of freedom estimation are also included. See
\code{pbkrtest} function \code{vcovAdj} for more information.}
}}

