% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RecAM_sim_pb}
\alias{RecAM_sim_pb}
\title{Example Simulated Repeated Measures Pseudo-bulk scRNA-Seq Data}
\format{

}
\usage{
RecAM_sim_pb
}
\description{
Sample \code{\link[SummarizedExperiment]{SummarizedExperiment}} object
containing pseudo-bulk data aggregated from scRNA-seq data simulated using
the \code{RESCUE} package.
}
\keyword{datasets}
