% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RecAM_sim_sce}
\alias{RecAM_sim_sce}
\title{Example Simulated Repeated Measures scRNA-Seq Data}
\format{
\code{\link[SingleCellExperiment]{SingleCellExperiment}} object with
the following data in object slots:

\describe{
\item{\code{counts}}{Matrix of raw counts with genes represented by rows and
cells represented by columns.}
\item{\code{normcounts}}{Data normalized using the variance stabilizing
transformation from the \code{sctransform} package.}
    \item{\code{colData}}{
        \describe{
            \item{sampleID}{Sample identifier}
            \item{subjectID}{Subject identifier}
            \item{time}{Timepoint identifier}
            \item{group}{Group identifier}
        }
    }
}
}
\usage{
RecAM_sim_sce
}
\description{
Sample \code{\link[SingleCellExperiment]{SingleCellExperiment}} object with
gene expression data simulated using the \code{RESCUE} R package.
}
\keyword{datasets}
