% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REBEL-package.R
\docType{package}
\name{REBEL-package}
\alias{REBEL}
\alias{REBEL-package}
\title{REBEL: Repeated Measures Empirical Bayes Differential Expression Using LMM}
\description{
Simulate data from repeated measures scRNA-seq data and analyze differential expression between groups/timepoints in individual cell types
}
\keyword{internal}
