% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebelLMM.R
\name{rebelLMM}
\alias{rebelLMM}
\title{Title}
\usage{
rebelLMM(
  fixedEffects,
  normalizedCounts = NULL,
  colData = NULL,
  pseudoBulk,
  subjectVariable,
  sampleVariable = NULL,
  REML = T,
  parallel = F,
  cores = 2,
  outputFits = F
)
}
\arguments{
\item{outputFits}{}
}
\description{
Title
}
